+++ modules/katamari/src/main/ts/ephox/katamari/api/Singleton.ts
@@ -2,19 +2,20 @@ import { Cell } from './Cell';
 import { Optional } from './Optional';
 
 interface Singleton<T> {
-  clear: () => void;
-  isSet: () => boolean;
-  set: (value: T) => void;
+  readonly clear: () => void;
+  readonly isSet: () => boolean;
+  readonly get: () => Optional<T>;
+  readonly set: (value: T) => void;
 }
 
 export interface Revocable<T> extends Singleton<T> { }
 
 export interface Api<T> extends Singleton<T> {
-  run: (fn: (data: T) => void) => void;
+  readonly run: (fn: (data: T) => void) => void;
 }
 
 export interface Value<T> extends Singleton<T> {
-  on: (fn: (data: T) => void) => void;
+  readonly on: (fn: (data: T) => void) => void;
 }
 
 const revocable = <T> (doRevoke: (data: T) => void): Singleton<T> => {
@@ -29,6 +30,8 @@ const revocable = <T> (doRevoke: (data: T) => void): Singleton<T> => {
 
   const isSet = () => subject.get().isSome();
 
+  const get = (): Optional<T> => subject.get();
+
   const set = (s: T) => {
     revoke();
     subject.set(Optional.some(s));
@@ -37,6 +40,7 @@ const revocable = <T> (doRevoke: (data: T) => void): Singleton<T> => {
   return {
     clear,
     isSet,
+    get,
     set
   };
 };
@@ -55,6 +59,8 @@ export const api = <T extends { destroy: () => void }> (): Api<T> => {
     subject.set(Optional.none());
   };
 
+  const get = (): Optional<T> => subject.get();
+
   const set = (s: T) => {
     revoke();
     subject.set(Optional.some(s));
@@ -68,6 +74,7 @@ export const api = <T extends { destroy: () => void }> (): Api<T> => {
     clear,
     isSet,
     set,
+    get,
     run
   };
 };
@@ -77,6 +84,8 @@ export const value = <T> (): Value<T> => {
 
   const clear = () => subject.set(Optional.none());
 
+  const get = (): Optional<T> => subject.get();
+
   const set = (s: T) => subject.set(Optional.some(s));
 
   const isSet = () => subject.get().isSome();
@@ -85,6 +94,7 @@ export const value = <T> (): Value<T> => {
 
   return {
     clear,
+    get,
     set,
     isSet,
     on